backend_tienda_vinos/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pydantic==2.9.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.9.0